chore(build): provide esm support for bundlers (#1197)
In this diff I added esm bundle which will allow user bundlers
to produce much smaller code. Currently regular babel output is
wrapped with module iife.

This project has a lot of files so with single esm file users will
bundle it slightly faster.

As a bonus we see the size of the project without dependencies.